Add basic undo/redo functionality
This works by simply keeping backups the entire relevant state of the
Wobbly project after each action and rolling back to such states on
undo/redo. This is feasible since Wobbly's internal state is not very
memory-intensive.

A drawback is that all list or table models have to be repopulated from
scratch, so selections in dialogs are cleared on every undo/redo. This
could be improved later on by saving the UI state before a rollback and
restoring it after it is completed.

Fixes dubhater/Wobbly#15 .